Added the ability to choose own Username and IP
I've reworked the 'def transferBackup()' function to allow the user to input their own target Username and IP Address.

This allows the code to not be reliant on hard coding, and allows the user to choose where they want to rsync their file to.

This is something I wanted to add in order to make it easier for others, and I figured out that I can just add variables and strings together to form one, long, cohesive input. Positively making progress with my python3 skills!