Added failsafe checks to terminal commands
I've learned that 'subprocess.run' has a way of checking the Exit Status codes of terminal commands being ran, so, I've added them to every single one that matters so as for the program to not spit out that the command ran successfully, even if it didn't.

Previously it just assumed that the command executed correctly without errors, and would allow you to process, without alerting you of any issues, but now it alerts you of issues and as to why they've occurred.

I've also learned that I can import certain features that I need, and call those from it, instead of always having to 'subprocess.run([''])'. I can now just 'run([''])' instead.