Attempt to restore mySQL DB
I've attempted to add a function that will ask the user if they have created a mySQL / Maria Database.

If they say yes, it will ask for their username and password, then restore from a predefined location. # Change this if you want to use this yourself.

If they say no, it will ask them if they want to set it up, then ask for the username and password in order to backup the Database.sql file, change this as well if you want to use it.