Fixed mariaDB restoration not working.
It seems that I was trying to do '-u USERNAME -p PASSWORD' to restore the DB backup, but turns out that '-p PASSWORD' throws an error of the DB of 'PASSWORD' not existing.

The fix for this, is to skip '-p PASSWORD' entirely. I confirmed this to be working by using the 'msqldump' command and verifying the correct '.sql' file sizes.